% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_table_root}
\alias{collapse_table_root}
\title{Collapse time series table root. All keys listed below are known table root keys.}
\usage{
collapse_table_root(table, entry, pc)
}
\arguments{
\item{list}{entry: Time series entry}

\item{char}{pc: paleoData or chronData}
}
\value{
list table: Metadata
}
\description{
Collapse time series table root. All keys listed below are known table root keys.
}
