% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{download_from_url}
\alias{download_from_url}
\title{Download a LiPD file from a URL}
\usage{
download_from_url(path)
}
\arguments{
\item{path}{URL path to LiPD file}
}
\value{
path Local path to downloaded file
}
\description{
Download a LiPD file from a URL
}
\keyword{internal}
