% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_num.R
\name{export_model}
\alias{export_model}
\title{Index model tables}
\usage{
export_model(models)
}
\arguments{
\item{list}{models: Metadata}
}
\value{
list models: Metadata
}
\description{
Index model tables
}
\keyword{internal}
