% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{extractTs}
\alias{extractTs}
\title{Extract time series from LiPD data}
\usage{
extractTs(D, whichtables = "all", mode = "paleo")
}
\arguments{
\item{D}{LiPD data, sorted by dataset name : list}

\item{whichtables}{: char: Options: "all", "summ", "meas", "ens" : Table type to output in the time series.}

\item{mode}{: char: Options: "paleo", "chron"}
}
\value{
ts:  Time series : list
}
\description{
Create a time series from a library of LiPD datasets. A Time series is a flattened version of LiPD data that can be queried and filtered for easier data analysis.
}
\examples{
D <- readLipds()
ts <- extractTs(D)

}
\author{
Chris Heiser
}
