% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{find_data_dir}
\alias{find_data_dir}
\title{Use a recursive file search to find the "data" directory of a LiPD file}
\usage{
find_data_dir()
}
\arguments{
\item{char}{x: File type}
}
\value{
char files: Matching file paths
}
\description{
Use a recursive file search to find the "data" directory of a LiPD file
}
\keyword{internal}
