% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{fix_doi}
\alias{fix_doi}
\title{Switch DOI from BibJSON structure 'identifier' key to a root level "doi" key}
\usage{
fix_doi(d)
}
\arguments{
\item{list}{d: Metadata}
}
\value{
list d: Metadata
}
\description{
Switch DOI from BibJSON structure 'identifier' key to a root level "doi" key
}
\keyword{internal}
