% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_vn}
\alias{get_vn}
\title{Get dataSetName from metadata. If one is not found, use filename as fallback.}
\usage{
get_vn(vn, names)
}
\arguments{
\item{char}{vn: Variable name}
}
\value{
char vn: Variable name, possibly appended with a number
}
\description{
Get dataSetName from metadata. If one is not found, use filename as fallback.
}
\keyword{internal}
