% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hasData}
\alias{hasData}
\title{Check if metadata path exists. Combine path and i to check for existence}
\usage{
hasData(path, i)
}
\arguments{
\item{path}{Path in metadata}

\item{i}{Next path level.}
}
\value{
dat Data found or null
}
\description{
Check if metadata path exists. Combine path and i to check for existence
}
\keyword{internal}
