% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_num.R
\name{idx_table_by_num}
\alias{idx_table_by_num}
\title{Index tables in a loop}
\usage{
idx_table_by_num(tables)
}
\arguments{
\item{list}{tables: Metadata}
}
\value{
list tables: Metadata
}
\description{
Index tables in a loop
}
\keyword{internal}
