% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_name.R
\name{import_model}
\alias{import_model}
\title{Index model tables}
\usage{
import_model(models)
}
\arguments{
\item{list}{models: Metadata}
}
\value{
list models: Metadata
}
\description{
Index model tables
}
\keyword{internal}
