% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_name.R
\name{import_section}
\alias{import_section}
\title{Change index-by-number for one section}
\usage{
import_section(section)
}
\arguments{
\item{list}{section: Metadata}
}
\value{
list section: Metadata
}
\description{
Change index-by-number for one section
}
\keyword{internal}
