% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.url}
\alias{is.url}
\title{Check if given string is a URL or not}
\usage{
is.url(x)
}
\arguments{
\item{x}{Path or URL string}
}
\value{
bool
}
\description{
Check if given string is a URL or not
}
\keyword{internal}
