% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_metadata}
\alias{merge_csv_metadata}
\title{Using the given metadata dictionary, retrieve CSV data from CSV files, and insert the CSV
values into their respective metadata columns.}
\usage{
merge_csv_metadata(d)
}
\arguments{
\item{list}{d: Metadata}
}
\value{
list d: Metadata
}
\description{
Using the given metadata dictionary, retrieve CSV data from CSV files, and insert the CSV
values into their respective metadata columns.
}
\keyword{internal}
