% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{merge_interpretations}
\alias{merge_interpretations}
\title{Merge the old interpretation fields into the new, combined interpretation field}
\usage{
merge_interpretations(d, key)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{key: climateInterpretation or isotopeInterpretation}
}
\value{
list d: Metadata
}
\description{
Merge the old interpretation fields into the new, combined interpretation field
}
\keyword{internal}
