% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{process_expression}
\alias{process_expression}
\title{Worker function that sends through the query expressions for finding matching entries in the time series}
\usage{
process_expression(ts, expression, exact = FALSE)
}
\arguments{
\item{ts}{Time series : list , Time Series data}

\item{expression}{Search expression : char (single query) or list (multiple query)}

\item{exact}{Key match : char. Is the provided key an exact key match or a piece of the key? ie. paleoData_variableName or variableName?}
}
\value{
results: list , Contains list of matching time series entries and a list of matching time series indicies.
}
\description{
Worker function that sends through the query expressions for finding matching entries in the time series
}
\author{
Chris Heiser
}
