% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_names_from_lists}
\alias{remove_names_from_lists}
\title{remove names from lists}
\usage{
remove_names_from_lists(L)
}
\arguments{
\item{L}{}
}
\value{
L without named lists in key spots
}
\description{
remove names from lists
}
