% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2.R
\name{ts2tibble}
\alias{ts2tibble}
\title{Convert a LiPD TS object into an equivalent nested tibble}
\usage{
ts2tibble(TS)
}
\arguments{
\item{TS}{}
}
\value{
a nested tibble
}
\description{
Convert a LiPD TS object into an equivalent nested tibble
}
