% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2.R
\name{untidyTs}
\alias{untidyTs}
\title{create tidy data.frame from TS}
\usage{
untidyTs(tTS, age.var = "age")
}
\arguments{
\item{tTS}{a tidy data frame, such as those created by tidyTs()}

\item{age.var}{}
}
\value{
a LiPD Timeseries object
}
\description{
takes a long, tidy, data.frame and returns a TS object. This is the opposite of tidyTs().
}
\seealso{
Other LiPD manipulation: 
\code{\link{pullTsVariable}()},
\code{\link{pushTsVariable}()},
\code{\link{tidyTsOld}()},
\code{\link{tidyTs}()}
}
\concept{LiPD manipulation}
