% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zips.R
\name{unzipper}
\alias{unzipper}
\title{Unzip LiPD file to the temporary directory}
\usage{
unzipper(path, dir_tmp)
}
\arguments{
\item{char}{tmp: Temporary directory}
}
\value{
none
}
\description{
Unzip LiPD file to the temporary directory
}
\keyword{internal}
