% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{update_lipd_v1_3}
\alias{update_lipd_v1_3}
\title{Update LiPD v1.2 to v1.3
- Added 'createdBy' key
- Top-level folder inside LiPD archives are named "bag". (No longer <datasetname>)
- .jsonld file is now generically named 'metadata.jsonld' (No longer <datasetname>.lpd )
- All "paleo" and "chron" prefixes are removed from "paleoMeasurementTable", "paleoModel", etc.
- Merge isotopeInterpretation and climateInterpretation into "interpretation" block
- ensemble table entry is a list that allows multiple tables
- summary table entry is a list that allows multiple tables}
\usage{
update_lipd_v1_3(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Metadata
}
\description{
Update LiPD v1.2 to v1.3
- Added 'createdBy' key
- Top-level folder inside LiPD archives are named "bag". (No longer <datasetname>)
- .jsonld file is now generically named 'metadata.jsonld' (No longer <datasetname>.lpd )
- All "paleo" and "chron" prefixes are removed from "paleoMeasurementTable", "paleoModel", etc.
- Merge isotopeInterpretation and climateInterpretation into "interpretation" block
- ensemble table entry is a list that allows multiple tables
- summary table entry is a list that allows multiple tables
}
\keyword{internal}
