% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{update_lipd_v1_3_keys}
\alias{update_lipd_v1_3_keys}
\title{Update v1.2 keys to v1.3 keys: recursive}
\usage{
update_lipd_v1_3_keys(d)
}
\arguments{
\item{list}{d: Metadata}
}
\value{
list d: Metadata
}
\description{
Update v1.2 keys to v1.3 keys: recursive
}
\keyword{internal}
